Malloc uses a 32-bit value for size on 32-bit systems

+++ common/runtime/ct_runtime.h
@@ -88,7 +88,7 @@ void __ctStoreThreadCreate(unsigned int, long long, ct_tsc_t);
 void __ctStoreThreadJoin(pthread_t, ct_tsc_t);
 void __ctStoreSync(void*, int, int, ct_tsc_t);
 void __ctStoreBarrier(bool, void*, ct_tsc_t);
-void __ctStoreMemoryEvent(bool, unsigned long long, void*);
+void __ctStoreMemoryEvent(bool, size_t, void*);
 void* __ctInitThread(void*);//pcontech_thread_create ptc
 void __ctCheckBufferSize(unsigned int);
 void __ctStoreDelay(ct_tsc_t start_t);
